Add DashSlideContainer Actor.
A simple Actor with one child whose allocation takes into account the slide out
of its child via the _slidex parameter ([0:1]).

This is required since I want to track the input region of this container which
is based on its allocation even if the child overlows the parent actor.

This will solve the problem with multimonitor where the slided out part of the
dash even if invisible due to the clipping still steals input events. Hopefully
it will also help to simplify the code a bit.