Use DashSliderContainer. Code cleanup.
Start to use the new DashSliderContainer. This should solve the bug concernig
stolen clicks with multimonitor-configuration and primary monitor on right. The
code has been partially cleanup with reorganization of the dash constructor.
Some code is necessary anymore with the new container such as the clipping
function and the Clutter.Gravity machinery and thus has been removed.